+++ changelog.md
@@ -1,3 +1,6 @@
+### [20-10-2020]
+* static config extraction lookup in database before scan file with yara and extract
+
 ### [15-10-2020]
 * Huge code unification and cleanup between `submission/views.py` and `api/views.py`
 * Improve error messages on bulk submission, for failed samples/hashes